+++ packages/web-core/src/services/inworld_connection.service.ts
@@ -17,6 +17,7 @@ import { InworldPlayer } from '../components/sound/inworld_player';
 import { InworldRecorder } from '../components/sound/inworld_recorder';
 import { Character } from '../entities/character.entity';
 import { InworldPacket } from '../entities/packets/inworld_packet.entity';
+import { EventFactory } from '../factories/event';
 import { ConnectionService } from './connection.service';
 
 interface InworldConnectionServiceProps<
@@ -225,6 +226,16 @@ export class InworldConnectionService<
         .getEventFactory()
         .narratedAction(text, params?.characters),
     );
+  }
+
+  async changeScene(name: string) {
+    this.connection.setNextSceneName(name);
+
+    return this.connection.send(() => EventFactory.loadScene(name));
+  }
+
+  async addCharacters(names: string[]) {
+    return this.connection.send(() => EventFactory.loadCharacters(names));
   }
 
   async sendCustomPacket(getPacket: () => ProtoPacket) {